# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PetConnect API
  description: |
    API REST para PetConnect — plataforma de mascotas que conecta dueños,
    veterinarios y comunidad. Incluye feed social, ranking semanal, InstaPet
    (red social por mascota), mascotas perdidas y adopciones.
  version: 0.1.0
  contact:
    name: PetConnect Team
servers:
  - url: http://localhost:8001/api/v1
    description: Servidor local de desarrollo (Hono)
  - url: https://api.petconnect.app/api/v1
    description: Producción
security:
  - bearerAuth: []
paths:
...
  /feed:
    get:
      tags: [Feed]
      summary: Obtener feed global de posts
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Posts paginados del feed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedPosts"
    post:
      tags: [Feed]
      summary: Crear nuevo post en el feed
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostRequest"
      responses:
        "201":
          description: Post creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "422":
          $ref: "#/components/responses/ValidationError"
  /feed/{post_id}:
    get:
      tags: [Feed]
      summary: Obtener detalle de un post
      parameters:
        - $ref: "#/components/parameters/PostId"
      responses:
        "200":
          description: Detalle del post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostDetail"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Feed]
      summary: Eliminar post
      parameters:
        - $ref: "#/components/parameters/PostId"
      responses:
        "204":
          description: Post eliminado
        "403":
          $ref: "#/components/responses/Forbidden"
  /feed/{post_id}/like:
    post:
      tags: [Feed]
      summary: Dar like a un post
      parameters:
        - $ref: "#/components/parameters/PostId"
      responses:
        "201":
          description: Like registrado
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      tags: [Feed]
      summary: Quitar like de un post
      parameters:
        - $ref: "#/components/parameters/PostId"
      responses:
        "204":
          description: Like removido
  /feed/{post_id}/comments:
    get:
      tags: [Feed]
      summary: Listar comentarios de un post
      parameters:
        - $ref: "#/components/parameters/PostId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Comentarios paginados
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedComments"
    post:
      tags: [Feed]
      summary: Comentar en un post
      parameters:
        - $ref: "#/components/parameters/PostId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentRequest"
      responses:
        "201":
          description: Comentario creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "422":
          $ref: "#/components/responses/ValidationError"
  /feed/{post_id}/comments/{comment_id}:
    delete:
      tags: [Feed]
      summary: Eliminar comentario
      parameters:
        - $ref: "#/components/parameters/PostId"
        - $ref: "#/components/parameters/CommentId"
      responses:
        "204":
          description: Comentario eliminado
        "403":
          $ref: "#/components/responses/Forbidden"

  # =========================================================================
  # RANKING
  # =========================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT de Supabase Auth

  # ------------------------------------------------------------------
  # Parameters (reusables)
  # ------------------------------------------------------------------
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
      description: Número de página
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
      description: Resultados por página
...
    PostId:
      name: post_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    CommentId:
      name: comment_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Forbidden:
      description: No autorizado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Conflicto (registro duplicado)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: Error de validación
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  # ------------------------------------------------------------------
  # Schemas
  # ------------------------------------------------------------------
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          example: "Recurso no encontrado"
      required: [detail]
    PaginatedResponse:
      type: object
      properties:
        items:
          type: array
          items: {}
        total:
          type: integer
          example: 42
        page:
          type: integer
          example: 1
        pages:
          type: integer
          example: 3

    # --- Auth ---
...
    Profile:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        user_id:
          type: string
          format: uuid
        username:
          type: string
          example: "petlover"
        avatar_url:
          type: string
          format: uri
          nullable: true
        bio:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # --- Pet ---
    Pet:
      type: object
      properties:
        id:
          type: string
          format: uuid
        owner_id:
          type: string
          format: uuid
        name:
          type: string
          example: "Luna"
        species:
          type: string
          example: "Perro"
        breed:
          type: string
          nullable: true
          example: "Golden Retriever"
        age:
          type: integer
          nullable: true
          example: 24
        weight:
          type: number
          nullable: true
          example: 25.5
        photo_url:
          type: string
          format: uri
          nullable: true
        bio:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Post:
      type: object
      properties:
        id:
          type: string
          format: uuid
        author_id:
          type: string
          format: uuid
        pet_id:
          type: string
          format: uuid
        content:
          type: string
          nullable: true
        photo_url:
          type: string
          format: uri
          nullable: true
        likes_count:
          type: integer
          example: 15
        comments_count:
          type: integer
          example: 3
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        author:
          $ref: "#/components/schemas/Profile"
        pet:
          $ref: "#/components/schemas/Pet"
    PostDetail:
      allOf:
        - $ref: "#/components/schemas/Post"
        - type: object
          properties:
            liked_by_me:
              type: boolean
              example: true
    CreatePostRequest:
      type: object
      properties:
        pet_id:
          type: string
          format: uuid
        content:
          type: string
        photo_url:
          type: string
          format: uri
      required: [pet_id]
    PaginatedPosts:
      allOf:
        - $ref: "#/components/schemas/PaginatedResponse"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Post"

    # --- Like ---
...
    Comment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        post_id:
          type: string
          format: uuid
        content:
          type: string
          example: "¡Qué linda foto!"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        author:
          $ref: "#/components/schemas/Profile"
    CreateCommentRequest:
      type: object
      properties:
        content:
          type: string
          example: "¡Qué linda foto!"
      required: [content]
    PaginatedComments:
      allOf:
        - $ref: "#/components/schemas/PaginatedResponse"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Comment"

    # --- Ranking ---
...